pyproject.toml
[project]
name = "speaker-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-speech>=2.33.0",
    "librosa>=0.11.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.1",
    "sounddevice>=0.5.2",
    "soundfile>=0.13.1",
    "streamlit>=1.50.0",
    "torchaudio>=2.8.0",
    "speechbrain>=0.5.16",
    "streamlit-webrtc>=0.47.6",
    "av>=14.4.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
]
